Breaking views.js Out
For parsability.
Using ES6 import/export syntax.
